pyproject.toml
[project]
name = "libris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "typer>=0.24.1",
]
[dependency-groups]
dev = [
  {include-group = "lint"},
  {include-group = "test"},
  "ipython"

]
lint = [
  "ruff"
]
test = [
  "pytest",
  "pytest-cov",
]

[project.scripts]
libris = "libris.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/libris"]
[tool.hatch.build.targets.wheel.sources]
"src" = ""
